requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn
cloudinary==1.44.0
django-import-export==4.3.7
django-filter==25.1
python-dotenv==1.1.0